refactor: move providers etc. to root to support proper Outlet slotting
